+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow infinite range in `fromRangeStep`
 - Panic instead of asserting where applicable
 - Restructured code
+- Rename `lazyOnce()` and `lazyRepeat()` to `onceWith()` and `repeatWith()` respectively
 
 ### Fixed